refactor: 重複したエラーハンドリングコードを共通化
- formatZodError関数をutils/error/に移動して共通化（97.47%の重複を解消）
- formatError関数も共通化し、ツール固有のデフォルト出力を引数で受け取る設計に変更
- calculateDamageハンドラー内のcreateCalculationResponse関数を追加して重複を削減
- 全テスト（231件）が成功することを確認
