+++ src/tools/calculateStatus/handlers/helpers/formatError.ts
@@ -1,76 +1,19 @@
-import { ZodError } from "zod";
+import { formatError as baseFormatError } from "@/utils/error";
 
-/**
- * ZodErrorを分かりやすいメッセージに変換する
- */
-const formatZodError = (error: ZodError): string => {
-  const issues = error.issues.map((issue) => {
-    const path = issue.path.join(".");
-
-    switch (issue.code) {
-      case "invalid_type":
-        if (issue.received === "undefined") {
-          return `「${path}」フィールドが必須です`;
-        }
-        return `「${path}」は${issue.expected}型である必要があります（現在: ${issue.received}型）`;
-
-      case "invalid_union":
-        return `「${path}」の形式が正しくありません`;
-
-      case "too_small":
-        return `「${path}」は${issue.minimum}以上である必要があります`;
-
-      case "too_big":
-        return `「${path}」は${issue.maximum}以下である必要があります`;
-
-      case "invalid_enum_value":
-        return `「${path}」は次のいずれかの値である必要があります: ${issue.options.join(", ")}`;
-
-      default:
-        return `「${path}」: ${issue.message}`;
-    }
-  });
-
-  return `入力エラー:\n${issues.join("\n")}`;
+// calculateStatus用のデフォルトエラー出力
+const DEFAULT_ERROR_OUTPUT = {
+  // エラー時でも最小限の情報を含める（スキーマ準拠のため）
+  pokemonName: "unknown",
+  stats: {
+    hp: 0,
+    atk: 0,
+    def: 0,
+    spa: 0,
+    spd: 0,
+    spe: 0,
+  },
 };
 
-export const formatError = (
-  error: unknown,
-): {
-  isError: true;
-  content: Array<{ type: "text"; text: string }>;
-} => {
-  const message = (() => {
-    if (error instanceof ZodError) {
-      return formatZodError(error);
-    }
-    if (error instanceof Error) {
-      return error.message;
-    }
-    return "不明なエラーが発生しました";
-  })();
-
-  const errorOutput = {
-    error: message,
-    // エラー時でも最小限の情報を含める（スキーマ準拠のため）
-    pokemonName: "unknown",
-    stats: {
-      hp: 0,
-      atk: 0,
-      def: 0,
-      spa: 0,
-      spd: 0,
-      spe: 0,
-    },
-  };
-
-  return {
-    isError: true,
-    content: [
-      {
-        type: "text",
-        text: JSON.stringify(errorOutput, null, 2),
-      },
-    ],
-  };
+export const formatError = (error: unknown) => {
+  return baseFormatError(error, DEFAULT_ERROR_OUTPUT);
 };
